* Fix an annoying warning

+++ src/game/g_local.h
@@ -1031,7 +1031,7 @@ void                G_DeletePTRConnection( connectionRecord_t *connection );
 extern  level_locals_t  level;
 extern  gentity_t       g_entities[ MAX_GENTITIES ];
 
-#define FOFS(x) ((int)&(((gentity_t *)0)->x))
+#define FOFS(x) ((size_t)&(((gentity_t *)0)->x))
 
 extern  vmCvar_t  g_dedicated;
 extern  vmCvar_t  g_cheats;
